requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
schedule==1.2.0
python-dotenv==1.0.0
feedparser==6.0.10
lxml==5.1.0
openai==1.12.0
